# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Healthy App",
        "description": "API documentation for Healthy",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/reviews": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get all reviews",
                "operationId": "346169ce6ab0c633530a7cedc1a25fab",
                "responses": {
                    "200": {
                        "description": "List of reviews",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Create a new review",
                "operationId": "154c71f5f72b459d998f5ea451bd4c96",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreReviewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Review created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/reviews/{id}": {
            "get": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Get a specific review",
                "operationId": "a9949d3a965ed93ae30547c07eef7505",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the review to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Review details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Update a review",
                "operationId": "17776becb9faf054b5c49bc83ad6c974",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the review to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateReviewRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Review updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reviews"
                ],
                "summary": "Delete a review",
                "operationId": "89986343b5dbd2b909c2764aeb7e4096",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the review to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Review deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get all users",
                "operationId": "c457726701591d1183b53aa71fc13441",
                "responses": {
                    "200": {
                        "description": "List of users",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UserResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Create a new user",
                "operationId": "592819a0265360b2014512d6dbfaf0e7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Get a specific user",
                "operationId": "36a33ff774d5cba33c039dec2c3e0287",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Update a user",
                "operationId": "b9091397c8b25f12c6adb74be6ce3a5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UserResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete a user",
                "operationId": "fa56cffde745d3f152f95cbacd936c0b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "User deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "StoreReviewRequest": {
                "required": [
                    "userId",
                    "mealId",
                    "rating"
                ],
                "properties": {
                    "userId": {
                        "description": "The ID of the user submitting the review",
                        "type": "integer",
                        "example": 1
                    },
                    "mealId": {
                        "description": "The ID of the meal being reviewed",
                        "type": "integer",
                        "example": 1
                    },
                    "rating": {
                        "description": "The rating given to the meal (1-5)",
                        "type": "integer",
                        "example": 5
                    },
                    "comment": {
                        "description": "An optional comment about the meal",
                        "type": "string",
                        "example": "This meal was delicious!"
                    }
                },
                "type": "object"
            },
...
            "StoreUserRequest": {
                "title": "Store User Request",
                "description": "Request body data for creating a new user",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "name": {
                        "description": "The name of the user",
                        "type": "string",
                        "maxLength": 255
                    },
                    "email": {
                        "description": "The email address of the user",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "The password of the user",
                        "type": "string",
                        "minLength": 8
                    },
                    "city": {
                        "description": "The city of the user",
                        "type": "string",
                        "maxLength": 255
                    },
                    "phone": {
                        "description": "The phone number of the user",
                        "type": "string",
                        "maxLength": 20
                    }
                },
                "type": "object"
            },
...
            "UpdateReviewRequest": {
                "properties": {
                    "userId": {
                        "description": "The ID of the user submitting the review",
                        "type": "integer",
                        "example": 1
                    },
                    "mealId": {
                        "description": "The ID of the meal being reviewed",
                        "type": "integer",
                        "example": 1
                    },
                    "rating": {
                        "description": "The rating given to the meal (1-5)",
                        "type": "integer",
                        "example": 5
                    },
                    "comment": {
                        "description": "An optional comment about the meal",
                        "type": "string",
                        "example": "This meal was delicious!"
                    }
                },
                "type": "object"
            },
...
            "UpdateUserRequest": {
                "title": "UpdateUserRequest",
                "description": "Request schema for updating a user",
                "properties": {
                    "name": {
                        "description": "The name of the user",
                        "type": "string",
                        "maxLength": 255
                    },
                    "email": {
                        "description": "The email of the user",
                        "type": "string",
                        "format": "email"
                    },
                    "password": {
                        "description": "The password of the user",
                        "type": "string",
                        "minLength": 8
                    },
                    "city": {
                        "description": "The city of the user",
                        "type": "string",
                        "maxLength": 255,
                        "nullable": true
                    },
                    "phone": {
                        "description": "The phone number of the user",
                        "type": "string",
                        "maxLength": 20,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AllergenResource": {
                "title": "Allergen Resource",
                "description": "Allergen resource representation",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the allergen",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the allergen",
                        "type": "string"
                    },
                    "mealsCount": {
                        "description": "The count of meals associated with the allergen",
                        "type": "integer"
                    },
                    "createdAt": {
                        "description": "The date when the allergen was created",
                        "type": "string",
                        "format": "date"
                    },
                    "updatedAt": {
                        "description": "The date when the allergen was last updated",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "IngredientResource": {
                "title": "Ingredient Resource",
                "description": "Ingredient resource representation",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the ingredient",
                        "type": "integer"
                    },
                    "name": {
                        "description": "The name of the ingredient",
                        "type": "string"
                    },
                    "nutritionalInfo": {
                        "description": "Nutritional information of the ingredient",
                        "properties": {
                            "calories": {
                                "description": "Calories per serving",
                                "type": "number",
                                "format": "float"
                            },
                            "macros": {
                                "description": "Macronutrient information",
                                "properties": {
                                    "protein": {
                                        "description": "Protein content in grams",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "carbohydrates": {
                                        "description": "Carbohydrate content in grams",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "sugar": {
                                        "description": "Sugar content in grams",
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "fat": {
                                        "description": "Fat content in grams",
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            },
                            "fiber": {
                                "description": "Fiber content in grams",
                                "type": "number",
                                "format": "float"
                            },
                            "sodium": {
                                "description": "Sodium content in milligrams",
                                "type": "number",
                                "format": "float"
                            },
                            "unit": {
                                "description": "Unit of measurement for the ingredient's nutritional values",
                                "type": "string"
                            }
                        },
                        "type": "object"
                    },
                    "mealsCount": {
                        "description": "Number of meals associated with this ingredient",
                        "type": "integer"
                    },
                    "mealQuantity": {
                        "description": "Quantity of the ingredient in the meal (if loaded via pivot)",
                        "type": "number",
                        "format": "float",
                        "nullable": true
                    },
                    "mealUnit": {
                        "description": "Unit of measurement for the ingredient in the meal (if loaded via pivot)",
                        "type": "string",
                        "nullable": true
                    },
                    "createdAt": {
                        "description": "The date when the ingredient was created",
                        "type": "string",
                        "format": "date"
                    },
                    "updatedAt": {
                        "description": "The date when the ingredient was last updated",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
            "MealResource": {
                "title": "Meal Resource",
                "description": "Meal resource representation",
                "properties": {
                    "id": {
                        "description": "The ID of the meal",
                        "type": "integer"
                    },
                    "title": {
                        "description": "The title of the meal",
                        "type": "string"
                    },
                    "description": {
                        "description": "The description of the meal",
                        "type": "string"
                    },
                    "price": {
                        "description": "The price of the meal in currency",
                        "type": "number",
                        "format": "float"
                    },
                    "isAvailable": {
                        "description": "Indicates if the meal is available",
                        "type": "boolean"
                    },
                    "availability": {
                        "description": "The availability time frame of the meal",
                        "properties": {
                            "from": {
                                "description": "The start time of availability",
                                "type": "string",
                                "format": "date-time"
                            },
                            "to": {
                                "description": "The end time of availability",
                                "type": "string",
                                "format": "date-time"
                            }
                        },
                        "type": "object"
                    },
                    "dietType": {
                        "description": "The diet type of the meal",
                        "type": "string"
                    },
                    "owner": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "allergens": {
                        "description": "List of allergens associated with the meal",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AllergenResource"
                        }
                    },
                    "ingredients": {
                        "description": "List of ingredients in the meal",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/IngredientResource"
                        }
                    },
                    "createdAt": {
                        "description": "The creation date of the meal",
                        "type": "string",
                        "format": "date"
                    },
                    "updatedAt": {
                        "description": "The last update date of the meal",
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            },
...
            "ReviewResource": {
                "title": "Review Resource",
                "description": "Review resource representation",
                "properties": {
                    "id": {
                        "description": "The unique identifier of the review",
                        "type": "integer"
                    },
                    "user": {
                        "$ref": "#/components/schemas/UserResource"
                    },
                    "meal": {
                        "$ref": "#/components/schemas/MealResource"
                    },
                    "rating": {
                        "description": "The rating given by the user",
                        "type": "integer"
                    },
                    "comment": {
                        "description": "The comment provided by the user",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "UserResource": {
                "title": "Allergen Resource",
                "description": "Allergen resource representation",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}